Add Test::Pod test dependency
After having built a pristine Perl with perlbrew it turned out that
after having installed all the dependencies via `cpanm --installdeps .`
that the `Test::Pod` module was still required in order to run `dzil
test`.  This patch adds `Test::Pod` to the list of test dependencies so
that this issue doesn't happen to someone else.