Generate "provides" info in META files
The experimental CPANTS metric `meta_yml_has_provides` tests if the
`provides` information exists in the META.yml and META.json files.  The
easiest way to implement this is in `Dist::Zilla` projects is to use the
`MetaProvides::Package` plugin, which automatically generates this
information and adds it to the META.* files.  This plugin has been added
to `dist.ini` and as a new dependency to the distribution; the relevant
"provides" metadata has also been added.